Polish 0.8.2 support
__ObjectiveDDP__

- Update Podspec
  - Update OS platform
  - Remove Libcrypto dependency
- Remove unused SRP methods
- Remove SRP specific login retry logic
- Fix broken specs
- Delete specs that are no longer required (SRP)
- Whitespace police (update style of PR to be more inline with existing code)

__Example App__

- Update LoginViewController to be compatible with new login method callback messages
- Update example meteor app (todos) to 0.8.2

+++ ObjectiveDDP/BSONIdGenerator.m
@@ -4,7 +4,7 @@
 
 static NSInteger methodCallCount = 1;
 + (NSString *)generate {
-  return [NSString stringWithFormat:@"%d", methodCallCount++];
+  return [NSString stringWithFormat:@"%ld", (long)methodCallCount++];
 }
 
 @end
